+++ RELEASES.md
@@ -39,6 +39,7 @@
   - Rounding errors may still occur when converting from arbitrary directions to the other 3592 discrete decidegrees.
 - `InputStreams` and `MutableInputStreams` no longer store e.g. `Option<Res<Input<MouseButton>>>`, and instead simply store `Res<Input<MouseButton>>`
   - This makes them much easier to work with and dramatically simplifies internal logic.
+  - added the `mockable_world` builder method to generate a `World` with the required resources.
 - `InputStreams::from_world` no longer requires `&mut World`, as it does not require mutable access to any resources.
 
 ## Version 0.4.1